Datatype adjustment

+++ imbalance_baselines/utils.py
@@ -10,6 +10,7 @@ def sanitize_str(inp, casttype):
 
 # TODO [3]: This weight function should be renamed/refactored to better represent the
 #   specific method it uses (effective no. of samples?)
+# TODO [4]: Pass precision preference through cfg
 def get_weights(class_sizes, beta=0, dtype: torch.dtype = torch.double,
                 device: torch.device = torch.device("cpu")) -> torch.Tensor:
     """Get normalized weight (inverse of effective number of samples) per class."""
@@ -33,8 +34,7 @@ def get_weights(class_sizes, beta=0, dtype: torch.dtype = torch.double,
     ).to(device)
 
     # Normalize the weights
-    # TODO [4]: Pass precision preference through cfg
-    weights = torch.mul(weights, class_cnt / torch.sum(weights)).double()
+    weights = torch.mul(weights, class_cnt / torch.sum(weights))
     
     weights.requires_grad = False
